Modularizes color styles export logic
Extracts color styles export functionality into a separate module for improved code organization and reusability.
Adjusts UI dimensions and updates references to the newly modularized function.
